RPRMAX-522: Project files and code fixes for 3ds Max 2020 (#184)
* RPRMAX-522: Project files and code fixes for 3ds Max 2020

Done.  Added project configurations Release-2020 and Debug-2020.

* Updated readme

* Added new .props files /PropertySheets/3dsmax-2020.  Copied from 3ds Max SDK under <maxsdk>\ProjectSettings\propertySheets, which should be the most compatible possible settings.  A few changes to the .props were needed, (1) changed $(Configuration) to $(MaxConfiguration) which is defined as "Release" or "Debug" and needed by other settings, instead of "Release-2020" and "Debug-2020", and (2) defined $(MaxSDK) using an environment variable instead of a relative path.

* Updated project file MaxPlugin.vcxproj, and enforced Platform Toolset version v141 for 3dsmax2020 build, and version v140 for 3dsmax2019 and earlier builds.

* Fixed compiler warnings.  Most changes for implicit typecasing warnings from double -> float.  This required changing hundreds of numeric constants in FireRender.Max.Plugin/plugin/SkyGen.cpp.  Also note the 3ds max property sheet now enforces floating point "precise" mode instead of "fast" mode for 3dsmax2020.

* Did not fix a compiler warning for redefinition of contant M_PI, in FireRender.Components/cpp/SunPosition/SPA.cpp.  The file originates from another depot and should not be changed here.

* Adding support for 2020 and removing 2017.

+++ FireRender.Max.Plugin/parser/SceneParser.cpp
@@ -1766,8 +1766,8 @@ void ApplyFREnvironment(frw::Scope& scope, frw::MaterialSystem& materialSystem, 
 				Matrix3 tx = tm;
 
 				// flip yz
-				tx.PreRotateX(M_PI_2);
-				tx.PreRotateY(M_PI);
+				tx.PreRotateX((float)M_PI_2);
+				tx.PreRotateY((float)M_PI);
 
 				ibl.SetTransform(tx);
 			}
@@ -2045,8 +2045,8 @@ bool SceneParser::parseEnvironment(Texmap* enviroMap)
 				angle = -sunAzimuth / 180.0f * M_PI - M_PI * 0.5;
 			}
 			// flip yz
-			tx.PreRotateX(M_PI_2);
-			tx.PreRotateY(M_PI);
+			tx.PreRotateX((float)M_PI_2);
+			tx.PreRotateY((float)M_PI);
 			if (angle != 0)
 				tx.PreRotateY(angle);
 			if (enviroImage)
@@ -2759,8 +2759,8 @@ void SceneParser::parseMaxLight(const ParsedNode& parsedNode, Object* evaluatedO
 						Matrix3 tx;
 						tx.IdentityMatrix();
 						// flip yz
-						tx.PreRotateX(M_PI_2);
-						tx.PreRotateY(M_PI);
+						tx.PreRotateX((float)M_PI_2);
+						tx.PreRotateY((float)M_PI);
 
 						tm = tx*r*tm;
 						}
